feat: use of User object - Django's model
This is User model with many useful methods from django, basically when
we create a user via 42api, we populate this object to access these
method, making our job easier
