+++ backend/auth_app/settings.py
@@ -42,6 +42,10 @@ INSTALLED_APPS = [
     # 'auth_app.middleware.updateLastActivity',
 ]
 
+AUTHENTICATION_BACKENDS = [
+    'django.contrib.auth.backends.ModelBackend',
+]
+
 MIDDLEWARE = [
     'django.middleware.security.SecurityMiddleware',
     'django.contrib.sessions.middleware.SessionMiddleware',
@@ -50,6 +54,7 @@ MIDDLEWARE = [
     'django.contrib.auth.middleware.AuthenticationMiddleware',
     'django.contrib.messages.middleware.MessageMiddleware',
     'django.middleware.clickjacking.XFrameOptionsMiddleware',
+    # 'django_pudb.PudbMiddleware',
 ]
 
 LOGIN_REDIRECT_URL = 'dashboard/'
