+++ backend/auth_app/templates/logout.html
@@ -7,5 +7,3 @@
 		{% endfor %}
 	</div>
 {% endif %}
-
-<p>sucker</p>
